backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
flask-restful
werkzeug==2.3.7
sqlalchemy==1.4.0
flask-cors
flask-security-too
flask-caching
celery[redis]